commit task2

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -8,7 +8,7 @@
  */
 int _strlen(char *s)
 {
-	size_t length = 0;
+	int length = 0;
 
 	while (*s++)
 		length++;
